// src/app/models/read/item.model.ts
export class Item {
    nazivproizvoda: string | undefined;
    opis: string | null = null;
    cijena: number | null = null;
    datumakcije: string | null = null;
    URLdoslikeUri: string | null = null;
    retailerId: number | null = null;
    nazivretailera: string | null = null;
    kategorija: number | null = null;


}